Move Actor and Film classes to Models directory.
