+++ Repository.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Data.SqlClient;
+﻿using ADOnetSakilaKoppling.Models;
+using Microsoft.Data.SqlClient;
 using System;
 using System.Collections.Generic;
 using System.Linq;
